moved Socket Service Interface to its folder
