+++ lib/features/shared/data/repository/socket_service.dart
@@ -4,7 +4,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:logging/logging.dart';
 import 'package:malawi_ride_share_app/services/socket_service/socket_constants.dart';
-import 'package:malawi_ride_share_app/services/socket_service/socket_service_interface.dart';
+import 'package:malawi_ride_share_app/features/shared/domain/repositories/socket_service_interface.dart';
 import 'package:socket_io_client/socket_io_client.dart' as io;
 
 abstract class SocketService implements SocketServiceInterface {
