+++ lib/features/shared/domain/repositories/socket_service_interface.dart
