+++ lib/services/socket_service/rider_socket_service.dart
@@ -1,4 +1,4 @@
-import 'package:malawi_ride_share_app/services/socket_service/socket_service.dart';
+import 'package:malawi_ride_share_app/features/shared/data/repository/socket_service.dart';
 
 class RiderSocketService extends SocketService {
   @override
